+++ lib/errors/errors.dart
@@ -1,0 +1,11 @@
+part of '../paynow.dart';
+
+class HashMismatchException implements Exception {
+  final String cause;
+  HashMismatchException(this.cause);
+}
+
+class ValueError implements Exception {
+  final String cause;
+  ValueError(this.cause);
+}